+++ CHANGELOG.md
@@ -7,3 +7,12 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [Unreleased]
 
 - Initial release
+
+## [0.0.6]
+- Added new command to copy the selected issues web link to clipboard
+
+## [0.0.7]
+- Changed api call that gets all the gitlab issues to grab 1000 issues
+- Changed api call that gets allt he gitlab issues to ignore issues tagged with 'Testing'.
+- Added a new command to clear the selected gitlab issue and clears your ingame status. 
+- Fixed webpack build error